начало задания
